tasks/create_table_for_json/requirements.txt
functions_framework==3.8.2
google-cloud-bigquery==3.31.0